+++ react/components/IframeUtils.js
@@ -1,13 +1,19 @@
 import PropTypes from 'prop-types'
 import Cookies from 'js-cookie'
 import axios from 'axios'
-let isPricingV2Active = window && window.localStorage && window.localStorage.getItem('routePriceSheetFromS3') || null
+let isPricingV2Active =
+  (window &&
+    window.localStorage &&
+    window.localStorage.getItem('routePriceSheetFromS3')) ||
+  null
 
 const getEnv = () => Cookies.get('vtex-commerce-env') || 'stable'
 
-const startLoading = () => window.postMessage({ action: { type: 'START_LOADING' } }, '*')
+const startLoading = () =>
+  window.postMessage({ action: { type: 'START_LOADING' } }, '*')
 
-const stopLoading = () => window.postMessage({ action: { type: 'STOP_LOADING' } }, '*')
+const stopLoading = () =>
+  window.postMessage({ action: { type: 'STOP_LOADING' } }, '*')
 
 const checkPricingVersion = async () => {
   // isPricingV2Active oneOf[true,false,null]
@@ -16,36 +22,36 @@ const checkPricingVersion = async () => {
   // null => request failed or localStorage is empty
   if (isPricingV2Active === null) {
     const res = await axios('/api/pricing/pvt/api-configuration')
-    isPricingV2Active = res.data && res.data.routePriceSheetFromS3 || null
+    isPricingV2Active = (res.data && res.data.routePriceSheetFromS3) || null
     localStorage.setItem('routePriceSheetFromS3', isPricingV2Active)
   }
 }
 
 const DELOREAN_REGISTRY = [
-  "billing",
-  "bridge",
-  "checkout",
-  "creditcontrol",
-  "fms",
-  "fms-picking",
-  "io-vtex-loader",
-  "license-manager",
-  "logistics",
-  "message-center",
-  "myaccount",
-  "payment-provider",
-  "pci-gateway",
-  "picking",
-  "portal",
-  "pricing",
-  "rnb",
-  "suggestions",
-  "suggestion",
-  "suggestions/catalog-mapping",
-  "suggestion/catalog-mapping",
-  "surveys",
-  "vtexid",
-  "vtable",
+  'billing',
+  'bridge',
+  'checkout',
+  'creditcontrol',
+  'fms',
+  'fms-picking',
+  'io-vtex-loader',
+  'license-manager',
+  'logistics',
+  'message-center',
+  'myaccount',
+  'payment-provider',
+  'pci-gateway',
+  'picking',
+  'portal',
+  'pricing',
+  'rnb',
+  'suggestions',
+  'suggestion',
+  'suggestions/catalog-mapping',
+  'suggestion/catalog-mapping',
+  'surveys',
+  'vtexid',
+  'vtable',
 ]
 
 function componentDidMount() {
@@ -71,6 +77,7 @@ function updateChildLocale(locale) {
 
 const contextTypes = {
   account: PropTypes.string,
+  workspace: PropTypes.string,
   culture: PropTypes.object,
   emitter: PropTypes.object,
   navigate: PropTypes.func,
@@ -80,9 +87,9 @@ const propTypes = {
   params: PropTypes.object,
 }
 
-const getLegacyHeaderTabs = pathName => {
+const getLegacyHeaderTabs = (pathName) => {
   let tabs, title
-  switch(pathName.toLowerCase()) {
+  switch (pathName.toLowerCase()) {
     // IMPORT & EXPORT SECTION
     case '/admin/site/relatorio_skus.aspx':
     case '/admin/site/produtoexportacaoimportacaoespecificacaov2.aspx':
@@ -99,14 +106,20 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('Relatorio_Skus.aspx'),
         },
         {
-          label: 'appframe.navigation.catalog.importExport.productSpecification',
+          label:
+            'appframe.navigation.catalog.importExport.productSpecification',
           path: '/admin/Site/ProdutoExportacaoImportacaoEspecificacaoV2.aspx',
-          active: pathName.includes('ProdutoExportacaoImportacaoEspecificacaoV2.aspx'),
+          active: pathName.includes(
+            'ProdutoExportacaoImportacaoEspecificacaoV2.aspx'
+          ),
         },
         {
           label: 'appframe.navigation.catalog.importExport.skuSpecification',
-          path: '/admin/Site/ProdutoExportacaoImportacaoEspecificacaoSKUV2.aspx',
-          active: pathName.includes('ProdutoExportacaoImportacaoEspecificacaoSKUV2.aspx'),
+          path:
+            '/admin/Site/ProdutoExportacaoImportacaoEspecificacaoSKUV2.aspx',
+          active: pathName.includes(
+            'ProdutoExportacaoImportacaoEspecificacaoSKUV2.aspx'
+          ),
         },
         {
           label: 'appframe.navigation.catalog.importExport.importImages',
@@ -119,9 +132,12 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('ProdutoImagemExportacao.aspx'),
         },
         {
-          label: 'appframe.navigation.catalog.importExport.exportProductreviews',
+          label:
+            'appframe.navigation.catalog.importExport.exportProductreviews',
           path: '/admin/Site/ProdutoExportacaoImportacaoAvaliacao.aspx',
-          active: pathName.includes('ProdutoExportacaoImportacaoAvaliacao.aspx'),
+          active: pathName.includes(
+            'ProdutoExportacaoImportacaoAvaliacao.aspx'
+          ),
         },
       ]
       if (!isPricingV2Active) {
@@ -131,7 +147,7 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('SkuTabelaValor.aspx'),
         })
       }
-    break
+      break
     // SETTINGS SECTION
     case '/admin/site/configform.aspx':
     case '/admin/site/configseocontents.aspx':
@@ -166,7 +182,7 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('GeographicRegion.aspx'),
         },
       ]
-    break
+      break
     // REPORTS SECTION
     case '/admin/site/relatorioindexacao.aspx':
     case '/admin/site/relatorio_seguranca.aspx':
@@ -213,7 +229,7 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('LinksRelatorios.aspx'),
         },
       ]
-    break
+      break
     // ATTACHMENT SECTION
     case '/admin/site/anexo.aspx':
     case '/admin/site/skuservicotipo.aspx':
@@ -242,70 +258,70 @@ const getLegacyHeaderTabs = pathName => {
           active: pathName.includes('SkuVincularValorServico.aspx'),
         },
       ]
-    break
+      break
     case '/admin/site/giftlisttype.aspx':
       title = 'appframe.navigation.cms.settings.listTypes'
       tabs = []
-    break
+      break
     case '/admin/site/categories.aspx':
     case '/admin/site/categoriaform.aspx':
       title = 'appframe.navigation.catalog.categories'
       tabs = []
-    break
+      break
     case '/admin/site/marca.aspx':
     case '/admin/site/marcaform.aspx':
       title = 'appframe.navigation.catalog.brands'
       tabs = []
-    break
+      break
     case '/admin/site/store.aspx':
       title = 'appframe.navigation.channel'
       tabs = []
-    break
+      break
     case '/admin/site/vale.aspx':
       title = 'appframe.navigation.vouchers'
       tabs = []
-    break
+      break
     case '/admin/site/seller.aspx':
       title = 'appframe.navigation.sellerManagement.title'
       tabs = []
-    break
+      break
     case '/admin/site/skuseller.aspx':
       title = 'appframe.navigation.SKUbinding'
       tabs = []
-    break
+      break
     case '/admin/site/produto.aspx':
       title = 'appframe.navigation.catalog.products'
       tabs = []
-    break
+      break
     case '/admin/site/xml.aspx':
       title = 'appframe.navigation.catalog.xml'
       tabs = []
-    break
+      break
     case '/admin/site/skucondicaocomercial.aspx':
       title = 'appframe.navigation.catalog.commercialCondition'
       tabs = []
-    break
+      break
     case '/admin/site/fornecedor.aspx':
       title = 'appframe.navigation.catalog.suppliers'
       tabs = []
-    break
+      break
     case '/admin/site/skukit.aspx':
       title = 'appframe.navigation.catalog.sku-bundle'
       tabs = []
-    break
+      break
     case '/admin/site/skukitform.aspx':
       title = 'appframe.navigation.catalog.manage-sku-bundle'
       tabs = []
-    break
+      break
     case '/admin/a':
     case '/admin/a/':
       title = 'appframe.navigation.cms.pageHeader'
       tabs = []
-    break
+      break
     default:
       title = 'appframe.navigation.catalog.title'
       tabs = []
-    break
+      break
   }
   return { tabs, title }
 }